requirements.txt
openai
langchain
pinecone-client
streamlit
python-dotenv
tiktoken
PyPDF2
